reusable tabs component and settings page tabs
